python-flask/requirements.txt
Flask
requests
Flask_pyoidc
python-dotenv
